services/asset_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
psycopg[binary]==3.1.18
psycopg-pool==3.1.18
pydantic[email]==1.10.14
python-multipart==0.0.9
PyJWT==2.8.0
jinja2==3.1.3
